Reorg: split output module, add InputSource trait, move factories
Split the monolithic output.rs (~978 lines) into a directory module with
separate files per sink implementation, reducing merge conflicts for
parallel development. Move build_output_sink() factory from pipeline_v2
to output/mod.rs for better cohesion. Add InputSource trait and FileInput
wrapper in input.rs, updating pipeline_v2 to use the trait abstraction
for extensibility to new input types. Apply cargo fmt to existing files.
